Fix IHA deployment
InstanceHA deployments use the ComputeInstanceHA role.
Via change I65c87f792683c493e5427dc55f381570310e9f56
"Fix composable role mapping for Compute" we are not matching
any text after Compute and so ComputeInstanceHA remains
unmatched and IHA deploys do not work. The reason for this
is that the regex will match the 'I' and the 'H' separately.

Let's simplify the regex by using a negative lookup
which is supported in python.re (as suggested by dciabrin):
>>> import re
>>> re.match(r'^Compute(?!HCI).*$','ComputeHCIfoo')
>>> re.match(r'^Compute(?!HCI).*$','ComputeHCI')
>>> re.match(r'^Compute(?!HCI).*$','ComputeInstanceHA')
<_sre.SRE_Match object; span=(0, 17), match='ComputeInstanceHA'>
>>> re.match(r'^Compute(?!HCI).*$','ComputeBar')
<_sre.SRE_Match object; span=(0, 10), match='ComputeBar'>

Change-Id: Id751b655e0887dc0501e3e88640c5923446880c0