WIP Fix up race when creating many vms by creating image pool explicitely first
With these steps (partly taken from oooq) we gurantee that we have an
image pool before calling virt-install. THe issue we are solving is that
virt-install implicitely creates one if it does not exist and it does so
in a racy way when multiple virt-install instances are spawned in
parallel.

Fixes: RHOSINFRA-3240
Change-Id: I2404726d1ec4bdeea21c972765c14e74994ac207