Fix pcs test for checking when pacemaker is configured
In RHEL8 the pcs status output text when pacemaker is not running has
changed and will now output:
[root@networker-0 ~]# pcs status
Error: error running crm_mon, is pacemaker running?
  Error: cluster is not available on this node

The grep being used is quite fragile, so let's just use
crm_mon to know if the cluster is up.

The other reason to switch to crm_mon is to
avoid https://bugzilla.redhat.com/show_bug.cgi?id=1830552
on RHEL 8.2

While we're at it we also clean up a couple of when statements
so that we do not get the following warnings any longer:

  [WARNING]: when statements should not include jinja2 templating delimiters
  such as {{ }} or {% %}. Found: '{{ install.deployment.files | basename }}' == 'virt'

Change-Id: I34c0b4307fedaef82a8f2635a77051b1799bbafe
Closes-Bug: RHOSINFRA-3263